try to shut down plugins gracefully before killing them  (#19603)
Currently when we're trying to shut down plugins (e.g. because someone
sends a `SIGINT` (Ctrl+C) to the CLI), we just send a `SIGKILL` to the
plugin to make it shutdown.

Since SIGKILL can't be intercepted, this means we're not giving
plugins a chance to clean up.  And since we are sending the signal to
the process group, if the plugin creates a subprocess that creates
further process groups, the SIGKILL doesn't necessarily get to all
processes.

For example, `dlv` the Go debugger launches the debugged process in a
different process group.  Sending a `SIGKILL` to that means that `dlv`
will shut down, but the process being debugged will continue running.
So that process is both being leaked, and also hangs the CLI.

Work around that by sending a `SIGINT` to the plugins first, trying to
shut them down gracefully.  Only if that doesn't work, send a
`SIGKILL` to finally kill the process.

(The first commit is a small refactor, so `InterruptChildren` is
available outside of the display code).

Fixes https://github.com/pulumi/pulumi/issues/19509