schema check and reserve pulumi keyword (#19277)
Python codegen breaks due to the pulumi key name being reserved. This
change is a first pass at adding various checks to the schema to fail
instead of generating the failing program.

Also added some diags that were not returned due to error. Error should
be checked but diags might be useful (as they are in these tests) so I
moved some things around a bit to return them

Fixes #18994

---------

Co-authored-by: Fraser Waters <fraser@pulumi.com>