Warn about empty config values (#19483)
Emit a warning for config values that are left blank/empty. In YAML
these are equivalent to `null` (alternatively written as `~`), but in
Pulumi’s config system they get treated as empty strings `””`.

We’d like to add support for null values in Pulumi config, but that’s a
change in behaviour. For now we only emit a warning, so that users are
aware of this ambiguity.

Note that when running `pulumi config set …` we load and write the
config out again. This already transforms any empty/`null`/`~` into an
empty string`””`.