retry downloads form upstream in `make ensure` (#19651)
In CI, `make ensure` sometimes fails because upstream servers return a
5xx or other random network issues. We already have `scripts/retry` to
retry things in CI, and these targets should retry just in case we hit a
temporary hiccup, e.g. such as in

https://github.com/pulumi/pulumi/actions/runs/15272119669/job/42950379398.

`scripts/retry` sets some extra environment variables that help with `go
test`, but they won't interfere with this, and it's convnient to have a
single script for retrying.