Permit providers to mark resource states as RefreshBeforeUpdate (#19487)
Permit providers to mark resource states as RefreshBeforeUpdate.

If the engine is processing such a state it will work as if the user
called `pulumi up --refresh` on updates.

Fixes https://github.com/pulumi/pulumi/issues/19303

Concretely:

- gRPC provider protocol is extended to permit marking the returned
state as RefreshBeforeUpdate in Create, Update, and
  Read method results
- State V3 grammar is extended to be able to save RefreshBeforeUpdate
flag alongside a resource state
- Plumbing and provider library wrappers are updated as necessary
- A test is added covering create, update, refresh and import behavior
for a provider that uses RefreshBeforeUpdate

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>
Co-authored-by: Will Jones <will@sacharissa.co.uk>