Add a conformance test for policy (#19631)
This adds a simple policy test to the conformance suite. 

Policy packs have to be manually written, and so languages tell the
conformance runner where they are via a new "PolicyPackDirectory" field
to `PrepareLanguageTests`. The conformance runner on hitting a test that
uses policy (which will all start with "policy-" as opposed to like
"l1-") will find the policy pack folder in the given directory, call
`InstallDependencies` on it to setup venvs/node_modules/etc then passes
it to the engine to start up like normal. The one small change needed
here is policy startup checks the existence of a language plugin to
decided on back-compat behaviour, it used to just do this via a normal
`GetPluginPath` lookup, but the language runtimes in conformance testing
aren't findable by that so we thread an override function down.

Policy validations are checked by looking for PolicyViolationEvent in
the engine events passed to the assert functions.

I'll follow this up with another test to check that policy config works
as expected, I've wired in the start of it in this PR to write out JSON
files with the config, but this first test doesn't actually use it to
check it works.

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>