run unit tests for all requested version sets (#19665)
Currently we run unit tests only for the "current" version set. This
seems mostly okay, as they don't do much with NodeJS or Python, however
for Go it's probably better if we do run the unit tests through all the
version sets as well. We have enough CI capacity to just do that without
worrying too much, so let's do it here.

This requires a small update for a test which behaves differently in Go
1.23 and 1.24.

Noticed this while looking into
https://github.com/pulumi/pulumi/pull/19653 (though it's probably less
important than that).