improve error handling in lifecycletest framework (#19678)
Small improvement to the lifecycletest framework. Currently if there's
an error, we log it and call `t.FailNow()` (and then unnecessarily still
assert that there's no error). Refactor this, so it's all a single
`require.NoError`, which in addition to being cleaner (and more
importantly) also provides the full backtrace for the error, which makes
it easier to debug.

Tiny thing I noticed while working with lifecycletests.