Use plugin.ExecPlugin in plugin run (#19677)
Rather than using `exec.Cmd` just re-use the execution logic we already
have in the plugin system for `plugin run`.

This has a difference that we're `io.Copy`ing the standard io streams
rather than assigning `os.Stdout/err` directly to the commands io
streams. Should be fine for the simple uses of `plugin run` we have, but
if we want to support rich tty style things here we'd need to be smarter
about this.