New eventual state import system (#19339)
This is an overhaul of the `import` resource option, which should solve
a number of issues we've historically had with it.

The old import option would `Read` the resource from the provider, then
check and diff the program inputs and the returned inputs to check that
there was no diff. This was because historically we couldn't run other
steps after the import step and so to achieve the engines goal of "final
state equals the programs declared goal" we had to error if anything
didn't match. If we didn't error we'd have a program declaring a goal
state of one thing, and an imported resource with a different state.

The new system takes advantage of our recent changes to make the step
generator asynchronously re-enterable. We can run one step, then go back
to the step generator and generate more steps. So that is now what
import does.

First we generate and run an import step. The import step no longer
calls Check or Diff, it simply reads from provider and returns that
state. We then use that imported state as the "old" state for the rest
of step generation to see if we need to do an update, or if there's no
diff we can just stop there (we could generate a SameStep, but that's
just extra noise so we elide it).