Add remote stack config (#19166)
Add CLI support for interacting with cloud stack configs - where stack
configs are managed in the cloud instead of as files in the local
project directory.

When selecting a stack which is using cloud stack config, it operates
exactly the same as normal except for having no local
`Pulumi.[stack].yaml` file.

Create a new project using cloud stack config using the (currently
hidden) flag `--use-esc-env`:

```shell
pulumi new <template> --remote-stack-config
```

Create a new stack with cloud config:

```shell
pulumi stack init <stack-name> --remote-config
```

Attempting to modify config via the CLI will result in an error message
directing the use of ESC commands directly:

```
$ pulumi config set myKey someValue
error: config set not supported for remote stack config: 
use `pulumi env set my-project/esc-test-stack pulumiConfig.myKey someValue`
```

If the `--config-file` flag is used to specify an explicit location of a
stack config file, this will be used in place of the cloud stack config,
but a warning will be displayed.

If a local file for a stack config exists but the stack is configured to
use cloud stack config then a warning will be printed that it has been
ignored and the cloud stack config is in use.

This change has no impact on DIY backends.

## Implementation Details

The cloud's stack object has been extended to have an optional "config"
field which, when set, indicates that this stack is using cloud stack
config. This field contains a required field called "environment" which
is a single ESC environment used to load the config map instead of a
local stack configuration file. Additionally, the encryption
configuration can also be written to the cloud stack config in case the
user is customising their encryption provider to something other than
Pulumi cloud.

Currently, the service only supports cloud stack config referencing the
latest version of an environment. This might be revised in the future.
This is to give us the option of being able to seamlessly map `config
set` and similar commands through to create a new version of the
referenced environment.

Previously, all stack config code assumes that it's reading and writing
to a local file on disk. Therefore we introduce an abstraction on the
backend `Stack` interface to:
1. Indicate if the stack has remote state.
2. Load and save the stack config remotely rather than as a local file.

Related issues:
- https://github.com/pulumi/pulumi/issues/9313
- https://github.com/pulumi/pulumi/issues/3776
- https://github.com/pulumi/pulumi/issues/19557