Allow clientLanguageRuntimeHost to use other standard language plugins (#19713)
Should fix the issues with policy pack install, and also fix the use of
shimless with inline programs. Neither of these are tested right now
because we hit a load of other issues trying to put automatic tests in
place for them. But this _should_ fix the use of language runtimes in
inline programs.