Initial support for Views (#19467)
This PR adds initial support for views, i.e. virtual child resources of
an owning resource. Note the `PULUMI_ENABLE_VIEWS_PREVIEW` envvar must
be set to `true` to turn on the feature (the feature flag will be
removed imminently).

Views are saved in the state with a `ViewOf` property, which is the URN
of the owning resource. The view itself is a child resource of the
owning resource.

Views are managed by a provider publishing view steps during CRUD
operations of the owning resource. `ResourceStepPre` events are emitted
for each view step published. The view steps are executed during
execution of owning resource step's `Apply` method.

Part of https://github.com/pulumi/pulumi/issues/19240