Fix `get-schema` to use the loaded plugin host
When running `pulumi package get-schema`, a plugin host is constructed
using the active Pulumi project where available. Later on, however, we
don't use this host to load plugins for retrieving the schema, instead
constructing a new one. This means that, if a `Pulumi.yaml` file defines
`packages`, these packages are not consulted by `pulumi package
get-schema` and their schema can't be retrieved. This change fixes that
by using the already-constructed `pctx.Host`.