Use new goal state for resource refresh (#19578)
Fixes https://github.com/pulumi/pulumi/issues/19561

When we added `refresh --run-program` even though we would have the new
URN on hand at the point we called refresh we had continued to just pass
in the old state, and so the provider would see the old URN, name and
type.

This PR updates the step generator to send the programs new goal state
for the refresh step, the old inputs and outputs are still sent.
This means the provider will see the new URN and we'll write the new URN
to the state file after the refresh.