Fix plan generation for imported resources (#19750)
Engine tests weren't running with plan generation turned on, but in the
CLI we always enable this. That missed that we could hit a panic for
imported resources because it hit an error that it's plan already
unexpectedly existed.

It would be worth our time ensuring engine tests always test this flag.