Fix issue when deleting/creating views during refresh (#19756)
This fixes a snapshot integrity error in a drift scenario where a child
view is gone during read and should be recreated. In this scenario
`pulumi up` is run on a stack with an existing owning resource marked
`RefreshBeforeUpdate` that has a child view resource. In `Read` for the
owning resource, a delete step for the child view is published. Then in
`Update` for the owning resource, a create step for the child is
published. Without this change, a snapshot integrity error happens due
to a duplicate child view resource in the snapshot.

The problem is that refresh steps for views weren't being persisted to
the snapshot in the case of `RefreshBeforeUpdate`, so the original child
view was not deleted from the snapshot. This change fixes that.

A regression test is included, as well as additional coverage for the
same scenario except using `pulumi up --refresh` instead of
`RefreshBeforeUpdate` (which already worked without errors).

No changelog since this feature is still behind a feature flag.

Fixes #19755