Experimental go policy SDK (#19670)
This is a bare bones, pretty much just enough to pass conformance tests,
policy SDK for Go. The most important part is it adds a new `Link`
method so we can use the new in-source version of the core SDK inside
tested policy packs. We need that for Go, Dotnet, and Java now, and if
we pull policy into the node and python libraries will need it for those
as well.

Everything new is tagged under an experimental package (enginex,
policyx) so we don't need to sweat the details on this much.