Test and fix providers not being cleanly shutdown in package add (#19769)
Fixes https://github.com/pulumi/pulumi/issues/19764

We were calling `Close` directly on the provider plugin client objects,
which did correctly close the provider down cleanly. But it didn't
remove it from the list of providers that the host object was tracking.
When we then asked the host object to also close it tried to close each
provider and would error that the provider had unexpectedly already
shutdown.

Small fix to just use `host.CloseProvider` instead of closing the
provider object directly.