run unit tests on all platforms (#19653)
We currently run unit tests on linux only. However we have enough
compute capacity to run them on all platforms, and it's safer to do so.
Do so.

Note that a bunch of tests don't succeed on Windows (and some don't
succeed on MacOS). For the most part I just excluded them from being run
on those platforms here, leaving the cleanup for follow up PRs. This
gets us to a place where *new* tests are going to be run on all
platforms, so we don't have more regressions, and we can clean up the
platform specific excludes as we go. Some excludes are using a
`TestMain` in the package, which is because most of the tests in that
package do not run successfully, and it was too much work to go through
test by test.