Respect `--generate-only` when `convert`ing local SDKs
When converting programs that use parameterized providers, `pulumi
convert` will generate local SDKs and link them into the generated
program. In the case of NodeJS, this involves modifying the generated
`package.json` file to point to the locally generated SDK sources.
Currently, we use the `npm add`, `pnpm add` or `yarn add` commands to
achieve this, depending on the package manager that has been chosen.

Unfortunately, the `add` subcommand is too enthusiastic -- as well as
modifying `package.json`, it will also install packages and create a
`node_modules` folder. When the `--generate-only` flag is passed to
convert, this is a bug -- the flag signifies that code is only to be
generated and not run in any way. This change fixes things so that we
use the `npm pkg set` and `pnpm pkg set` commands, which modify
`package.json` only without installing anything. When `--generate-only`
is not passed, we use the `add` commands as we do today.

The same problem exists for Python and its package managers
(specifically `poetry` and `uv`). Thankfully these have similar
alternative commands that we can lean on when we only want to set up
linkages without installing SDKs. This PR takes care of these cases too.

> [!NOTE]
> In the case of NodeJS/TypeScript, when `yarn` has been selected as the
> package manager, we use `npm`'s `pkg set` command, since `yarn` does
> not possess one of its own. This > should be fine since we only support
> Yarn Classic, whose recommended installation method requires `npm`.

Fixes #19616