Close SourceIterator (#19740)
We need to close the SourceIterator so it will cleanup the resource
monitor. This will cancel any outstanding requests on the resource
monitor, which ensures that the calls get written to the
`PULUMI_DEBUG_GRPC` log.

This required updated many of the lifecycle tests since now the
`monitor.RegisterResource` calls always returns. Tests previously
asserted `NoError` on these, but we never actually hit the assertion.