send informational output for policy pack installation to stderr
Typically stdout is used for content the user expects the command to
print, e.g. in `curl -v` the data that actually comes back from the
website, while stderr is used for additional informational output,
e.g. in `curl -v` all the HTTP information.

Whether policy packs are installed is information output that the user
doesn't necessarily need, so it should go to stderr.  This also makes
the `--json` flag e.g. in `pulumi preview` work as expected, as the
JSON output will be on stdout, and can be redirected to a file/piped
to further commands separately, while the informational output still
goes to stderr, and will be visible to the user, but won't interfere
with redirection of the main data stream.