Fix RefreshBeforeUpdate and views (#19821)
`RefreshBeforeUpdate` was originally implemented in the same way as
`pulumi refresh --run-program`, but this approach can result in snapshot
integrity errors when additional operations are peformed on the resource
after the refresh because of duplicate entries in the snapshot and
dependencies in the wrong order.

This change avoids these issues by making `RefreshBeforeUpdate` work
more like `pulumi up --refresh`, where resources marked
`RefreshBeforeUpdate` are refreshed first at the start of the update,
then the snapshot is rebuilt with the refreshed state, and then the rest
of the update proceeds.

Part of this change also fixes an issue where we weren't updating the
`oldViews` cache on `Deployment` after a `--refresh` (or now
`RefreshBeforeUpdate` refresh) operation.

With this change, all the module refresh tests in
https://github.com/pulumi/pulumi-terraform-module/pull/374 pass,
including `Test_replace_drift_deleted`.

No changelog since views hasn't shipped yet without the feature flag.

Fixes https://github.com/pulumi/pulumi/issues/19755