skip tests that are missing environment setup (#19827)
We have a few tests that need environment variables set up to run. These
fail in CI for either tests from external contributors, or dependabot
PRs. Skip these tests when the right environment setup is missing.

This is a little annoying to test, since it's not going to fail on any
of our PRs, and just submitting a PR from an external account doesn't
work either, as the tests don't run (see
https://github.com/pulumi/pulumi/pull/17192). So I just picked the
failing tests out from a dependabot PR. I might have missed some here,
but this is hopefully moving in the right direction.

Note that this doesn't solve anything for the automation API tests yet.

Also interestingly, it seems that in CI we do set the enviornment
variable in some cases, but to an empty value, which means we also need
to check for that, and can't just use the `ok` return from `LookupEnv`.