auto/nodejs: rename stack internally when calling rename()
Currently when using `stack.rename()` in the NodeJS automation API, we
rename the stack, but internally keep using the old name.  This means
any subsequent operations on that stack object will also use that old
name.

For the cloud backend this *kinda* works, as I think the cloud just
redirects from the old name to the new one, as long as there's no new
stack that took over the old name.  However for the filestate backend
this doesn't work at all, since the stack is no longer available.

Fix this by setting the name of the stack correctly internally as
well.

Also fix the test so it works when `PULUMI_ACCESS_TOKEN` is not set.
Note that there's a little oddity here, as for local backends we don't
have a summary event for the rename here.  Not sure how to deal with
that correctly.