Fix proto checksum order on macOS (#19840)
`sort` does not sort the same way for some characters on macOS compared
to Linux, depending on the locale. By setting the locale to `C` we
ensure we get the same order here.

This remove the need to manually re-order the protobuf `.checksum.txt`
file when updating protobufs on macOS.