Add policy-enforcement-config conformance test (#19845)
This is to test that the special "enforcementLevel" config is respected
by policy packs.

This adds a new feature to the conformance tests "RunsShareSource" so
that we can write out just one program for multiple runs. This is useful
when we want to test the same program against different configs.