[AutomationAPI/Python] Fix Parallelism (#19855)
When running an inline python program using the automation API,
deployments are limited to 16 concurrent actions.

We can notice the issue in this demo, which uses the automation API
(with an inline program) :
[https://asciinema.org/a/01HCUcmfWo27RUyuEJtnPoMHt](https://asciinema.org/a/01HCUcmfWo27RUyuEJtnPoMHt)
We can see that 16 resources are created (which takes 10 seconds), then
the 14 last resources are created. This results in waiting 20 seconds,
while it was expected to create all resources simultaneously.

This demo runs the exact same program but using the pulumi binary :
[https://asciinema.org/a/olWRUsZGfrVNZ4FdBzLUDqUxF](https://asciinema.org/a/olWRUsZGfrVNZ4FdBzLUDqUxF)

I found out that unlike the `sdk/python/cmd/pulumi-language-python-exec`
script, `the sdk/python/lib/pulumi/automation/_server.py` file does not
define a new ThreadPoolExecutor for the asyncio loop.