Very simple remediate policy test (#19856)
This is a very basic test that a remediate policy can run and change a
resource. While doing this I found an engine bug that the engine didn't
overwrite the PolicyPackVersion on remediation results like it did for
analyze results. This caused all python remediations to report their
version as 0.0.1.

---------

Co-authored-by: Julien <julien@caffeine.lu>