run all integration tests on all platforms (#18055)
We currently exclude some integration tests on windows and macos. This
is (probably) done as an optimization to save some compute time on these
platform. However we should have enough CI capacity to run these tests
everywhere, and it's better to spend that capacity, rather than spend
brain time to try and figure out what needs to be run cross platform and
what doesn't.

Note that this PR ends up skipping a few tests, that we still need to
fix. If people are happy with this direction, I'll open issues for those
(as they didn't seem straightforward to fix), and we can look into them.
We'd be no worse of than we currently are since these tests are not
being run in the first place right now, but at least going forward we'd
avoid introducing new test failures since everything would be running in
CI.