Skip registry lookup when `--file` is passed (#19859)
This addressed a bug that Pulumi's own insights team found when running:

```
pulumi plugin install resource some-provider ${VERSION} --file ${FILE}
```

The registry should not be consulted when `--file` has been passed,
since the user has located the provider for us. This is the same reason
why the registry shouldn't be consulted when `--server`
(pluginDownloadURL) is passed.