make lint steps depend on `ensure`
The lint steps have recently started to require the `biome` dependency
in sdk/nodejs.  This is not available unless one has run `make ensure`
or a command with similar effect (installing the biome dependency)
before.  Make the lint step depend on `ensure` to make sure it can
always successfully run.