Refactor applying engine operations to use one code path
Presently, Pulumi operations such as `Update()`, `Refresh()`, etc. are
exposed as methods on the `Backend` interface. While this might imply
that the implementations of these operations are for the most part
backend-specific, this is not completely true. In actuality, the
backend-specific parts of each operation are already fairly well
factored out into a set of interfaces (`SnapshotPersister`,
`BackendClient`, etc.) which are passed to the same core `engine.*`
functions. The remainder of the code in each method is largely the same
across backends -- managing events channels, making sure the display
renders correctly, and so on. This change is about pulling this code out
so that it can be shared, making the following arguments for doing so:

* This code is the same by definition, not by coincidence -- there
  _should_ only be one way for a Pulumi operation to proceed correctly,
  and this is defined by the engine. The backend is just one component
  that the engine (or some backend-agnostic piece of code) should be
  manipulating, not the other way around.

* Related to/evidencing the above, whenever we add a new feature to the
  engine (e.g. `--run-program`), we find ourselves threading it through
  both our backends with identical code changes. This is not
  particularly expensive, but it is error prone, and we do miss things
  occasionally.

* If we ever want to handle multi-stack deployments (hopefully in the
  near future), it will be beneficial to be able to reason about all
  operations with a single conceptual code path, as opposed to the
  current "lie" that operations are in some sense backend-dependent.

If you buy the argument, here's the approach stubbed out in this change:

* The `Backend` interface is now responsible for describing how to
  `Begin` a new `Application` (a new interface which encapsulates the
  current `b.apply` patterns). `Application` allows closure over
  backend-specific data structures, such as the backend's implementation
  of stack references or `engine.UpdateInfo`. It also owns hooks into
  the bits of an application such as creating a `SnapshotPersister` or
  `BackendClient`.

* A new `backend.ApplyStack` function becomes the one-stop shop for
  applying all Pulumi operations, using `BeginApply()` and the resulting
  `Application` to drive backend-specific components. The
  `Application`'s `End` method cleans things up when done.

* All "backend-specific" (but not actually) `b.apply`s can now be
  deleted and replaced with `backend.ApplyStack`.

With these modifications, we could actually go further and e.g. remove
the `apply` argument from `backend.PreviewThenPromptThenExecute` etc.
entirely. Moreover, we might be able to remove the `Update()` etc.
methods from `Backend` altogether later on. To minimise changes here and
ease review/increase confidence, however, we punt those to later work.