Allow DIY backend to remove stacks with just a stack resource (#19883)
This was noticed in dotnet testing:
https://github.com/pulumi/pulumi-dotnet/pull/650/commits/5547cc03dd70ecd500bab203739e86e99ba4fc33

Seems the service will allow the removal of a stack even if it still has
it's root stack resource in it (i.e. not cleaned up completly via
destroy). But the DIY backend would error that the stack still had
resources.

This adds a test to show both behave the same now, and fixes the DIY
backend to allow removal when there is only one resource which is the
root stack.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>