Fix arguments in pip/poetry errors (#19890)
Minor fix up noticed in community slack. The errors for these commands
say the command ran was `python list -v --format json` but it was in
fact `python -m pip list -v --format json`. This is because
`p.ModuleCommand` adds the extra `-m pip` arguments to the command args,
but when we build the error we use the original args rather than the
full ones.