Warn more aggressively for very old CLI versions
When the Pulumi CLI (and as a consequence, the Pulumi engine) falls
_way_ behind, it is almost guaranteed that users will have a subpar
experience due to the number of issues we fix (such as snapshot
integrity bugs) and improvements we ship at a regular cadence. This
change tweaks the update prompt in these cases to emphasize that the
running version is really very old and that updating is more advisory
than opportunistic. Presently, we consider "very old" to either be 2
major versions (unlikely in the near future, but a safe bet) or 24 minor
versions (which assuming ~1 release a week is 24 weeks = 6 months, which
feels reasonable).

Part of #19105