fix flaky TestExplicitDeleteBeforeReplace
This test is flaky because occasionally the steps are not being
generated in the same order.  In particular sometimes the Same step
happens before the Replace step.  I'm assuming this is because of the
recently introduced asynchronicity of the step generator.

Either way what we care about here is that all the steps are
happening, not necessarily that they are always in the exact same
order.  Steps do need a partial ordering, so snapshot integrity is
still sustained, but we already check for snapshot integrity at every
test step in the framework anyway.