unflake TestMultipleResourceDenyDefaultProviderLifecycle (#19893)
In this test we check that rpc error message after the resource monitor
is being shut down. There's some racyness in the shutdown, and depending
on which shutdown stage we're in we either get `Unavailable`, or
`Canceled`. For the purpose of this test it doesn't really matter which
one we're getting, so check for both to unflake this.

Fixes https://github.com/pulumi/pulumi/issues/19826