Fix refresh views with --run-program (#19877)
Originally reported in https://github.com/pulumi/pulumi/issues/19755,
and worked around via https://github.com/pulumi/pulumi/pull/19821 which
changed the tfmodule views to refresh before the program. This didn't
fix tfmodules (or other views) when used with `--refresh --run-program`.

This fixes views to write to the snapshot correctly when being
refreshed, this follows the same logic as normal refresh steps. That is
they can either be pre-program-run non-persisted steps, or persisted
steps being run during the program.

With this fix in place we should be able to move the RefreshBeforeUpdate
back to how it was, but I haven't done that in this PR.