fix flaky TestExplicitDeleteBeforeReplace (#19897)
This test is flaky because occasionally the steps are not being
generated in the same order. In particular sometimes the Same step
happens before the Replace step. I'm assuming this is because of the
recently introduced asynchronicity of the step generator.

Either way what we care about here is that all the steps are happening,
not necessarily that they are always in the exact same order. Steps do
need a partial ordering, so snapshot integrity is still sustained, but
we already check for snapshot integrity at every test step in the
framework anyway.

Potentially we could just replace `AssertSameSteps` with the new
function, but I'm not sure if ordering matters *somewhere*, so I went
for the minimally necessary change for now.

Fixes https://github.com/pulumi/pulumi/issues/19063