Allow plugin run to run a local plugin file directly (#19903)
Was going to use this for testing plugin run but it will also be useful
for allowing shimless plugins where the local path is a folder rather
than a file. That doesn't work quite yet because it needs some
supporting work I'm doing in another PR, but this change is part of it
and small enough to go in by itself.