Unify `UpdateInfo` implementations (#19918)
The `engine.UpdateInfo` type captures information about a Pulumi update
(where an update is a general term for a Pulumi operation, and not just
e.g. a `pulumi up`). Until now, it has been an interface exposing
information about the Pulumi project being operated on and the selected
target (stack, configuration, etc.), with implementations for both DIY
and Cloud (`httpstate`) backends.

It turns out that these implementations are in fact largely the same --
data structures carrying the relevant `Project` and `Target`, with some
backend-specific bits tacked on. Hoisting out these backend-specific
bits into the code that actually cares about them allows us to make
`UpdateInfo` a dumb struct instead. This reduces the amount of update
code that is actually backend-specific, and should make it simpler to
further unify the code paths for DIY and HTTP backends when it comes to
application (see #19866 for a first pass at this prior to this
refactoring). In turn, this should make it easier to think about
multi-stack updates later on, where a single backend-agnostic engine
should be able to juggle multiple backends simultaneously.