Fix parameterised packages in node mocks (#19917)
This tests and fixes the use of parameterised packages in Node mocks.

Firstly there is the actual fix, we need to set
`store.supportsParameterization` to true when using mocks.
Secondly, I've added tests to Node and Go and Python, to check that the
resource registration _actually_ runs.
Thirdly, I've added some more debug logs to the node codebase that I
found useful tracking down what was happening here.