Clean up verbatim text in CLI help
Presently, Pulumi CLI help text serves double duty: as well as providing
inline help for users of the CLI, it is rendered via Markdown and served
as part of the HTML documentation at pulumi.com/docs. Consequently, it
can be beneficial to e.g. call out verbatim parts of help such as
example commands using backticks to improve their web rendering. This
means a slight hit on the CLI presentation, but seems like a worthwhile
compromise in most cases. This change does just this for the public
`package *` commands.

Fixes #19920