Fix skipped resources not being sent for stack analysis (#19926)
This fixes a bug where when resources were removed from a program but
not targeted they would not be sent to AnalyzeStack.

They should have been because they were still part of the stack, and do
correctly stay in the snapshot. But because AnalyzeStack only works on
resources recorded into `deployment.news` and because there was no step
registered for these removed and skipped resources they were never
recorded.

Fix is to add these resources to deployment.news when we notice their
deletes are being skipped in GenerateDeletes.