Add resource hook protobuf definitions (#19746)
Add protobuf definition for `RegisterLifecycleHook`. This takes a
callback (and later will have options, like for example `async`) that
the engine can then later call on the language runtime.

`RegisterResouceRequest` takes an optional `LifecycleHooksBinding` which
binds individual hooks to lifecycle events of the resource.

The `LifecycleHookRequest` and `LifecycleHookResponse` types are used to
marshal the request and response of the hook callbacks.