Add `preview_refresh` for the Python Automation API (#19900)
In https://github.com/pulumi/pulumi/pull/19373, we added the
`previewOnly` flag to the `refresh` command. However, this raised some
issues, as commands in the automation API use the `pulumi stack history`
to find the results of a given command. Because a preview command
doesn't produce any new information in the `stack history`, we would
instead return the results of whatever the previous command was
(https://github.com/pulumi/pulumi/issues/19443).

Unfortunately, if we want the preview information, we have to make do
with what we can find in the event summary, but this doesn't give us
enough information to construct a full `RefreshResult`. For that reason,
this PR introduces a new command, `preview_refresh` (to return
everything we _can_ know, which is a `PreviewResult`), and deprecates
the `preview_only` flag on `refresh`.

This problem will also exist for `destroy --preview-only`, but I'll
address that in a follow-up PR, and make the equivalent changes to the
other automation APIs.