Fix a panic with --target, --refresh and provider upgrades (#19935)
Fixes https://github.com/pulumi/pulumi/issues/19879

If we hit the case where the provider resource wasn't registered we
should be able to look it up from the old state. This fixes the case
where a resource would need to be sent to analysis but wasn't registered
in the update pass and so its provider was also not registered.