Synthesize delete steps for views during preview (#19952)
A provider's `Delete` RPC is responsible for publishing delete steps for
any views of the resource to properly report any partial delete
failures. However, `Delete` is not called during preview, so views are
currently not getting delete steps which makes it look like only the
owning resource will be deleted during a `pulumi destroy` preview --
there is nothing in the display about any of the resource's views.

This change fixes this by synthesizing and publishing delete steps for
any views of a resource that is going to be deleted during a preview.
This way, the deletes for the views are displayed to the user.

Fixes #19929