Fix locks being held while running transforms (#19950)
Fixes https://github.com/pulumi/pulumi/issues/19904

This fixes an engine bug with locks being held. The engine was locking
to collect transforms to run and then running them within the lock as
well. This meant two resources couldn't run transforms at the same time.

In the case of #19904 it meant if the provider race'd and registered
after the stack registration (which is arguably another bug that this
can even happen) it would get stuck waiting to collect its transforms
because the resource was stuck inside it's transform application phase
waiting for the provider to register.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>