Restore hash values in assets and archives after transforms (#19960)
Fixes https://github.com/pulumi/pulumi/issues/17792

Not all SDKs maintain the 'hash' values of asset and archive objects in
their in-memory model. As such when running transforms even if the
transform is just the `id` function the asset and archive objects can
lose their hash values. Providers rely on these values for diff'ing so
this can lead to recurring diffs.

This fixes the engine to recompute hashes when reading values from
transforms, the same as we recompute them when reading the initial
values from the program.