Include Status service logs in PULUMI_DEBUG_GRPC=$PWD/logs.json (#19973)
Newly introduced Status service has methods such as PublishViewSteps
that did not trace themselves when the user
requested debug logs with PULUMI_DEBUG_GRPC. This is now fixed.

The problem was that the were two instances of interceptor initiated
against the same log file, which the comments
indicated would not work as expected. Now the instances are cached in a
global map and the comment restriction lifted.