Construct cancellation
When an error occurs in a resource created inside a Node.js component resource (and the component resource has an output that depends on that resource), we keep the construct GRPC call open, which prevents the shutdown of the source iterator, and causes the operation to hang.

Why does the construct call hang? When the "child" resource registration inside the component resource fails, we don't resolve its properties due to https://github.com/pulumi/pulumi/pull/13070. When we then try to serialise the output of the component resource, we hang forever.

This PR ensures that construct calls are canceled when the resource monitor shuts down.

Fixes https://github.com/pulumi/pulumi/issues/19905