replace assert.NoError with require.NoError
The differentce between the `assert.*` and `require.*` function is
that `assert` lets the test continue and `require` makes the test fail
immediately.

When an error occurs subsequent steps in a test are almost always
useless to keep running, as at best they will provide confusing
output, and at worst there's going to be a dependency on the actual
result, resulting in a panic in the test framework.

Convert all the `assert.NoError` calls to `require.NoError` calls,
which means aborting the test immediately.  This results in fewer
confusing errors while running the tests.

We should probably do the same with at least `assert.NonNil` and
`assert.Len`, as that also usually has similar issues with subsequent
tests that then either want to access the nil object, or assert an
element in the array that doesn't exist.